refactor: service interface and remove depricated ResponseEntity
