+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Genre;
-import ru.yandex.practicum.filmorate.service.GenreService;
+import ru.yandex.practicum.filmorate.service.GenreServiceImpl;
 
 import java.util.Collection;
 import java.util.Optional;
@@ -17,10 +17,10 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/genres")
 public class GenreController {
-    private final GenreService service;
+    private final GenreServiceImpl service;
 
     @Autowired
-    public GenreController(GenreService service) {
+    public GenreController(GenreServiceImpl service) {
         this.service = service;
     }
